requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
setuptools==75.8.0
sqlparse==0.5.3
wheel==0.45.1
